pyproject.toml
[project]
name = "decreto3eolico"
version = "0.1.0"
description = ""
authors = [
    {name = "Geremia Pompei",email = "56674278+GeremiaPompei@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "<3.14,>=3.11"
dependencies = [
    "numpy (>=2.2.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "tk (>=0.1.0,<0.2.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "python-docx (>=1.1.2,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pyinstaller = "^6.12.0"

